# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GenXcare API
  version: 1.0.0
  description: REST contract for the GenXcare modular monolith.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/tasks:
    get:
      tags: [Tasks]
      operationId: listAssignedTasks
      responses:
        "200": { $ref: "#/components/responses/List" }
        default: { $ref: "#/components/responses/Error" }
  /v1/seniors/{seniorId}/tasks:
    parameters: [{ $ref: "#/components/parameters/SeniorId" }]
    get:
      tags: [Tasks]
      operationId: listTasks
      parameters: [{ $ref: "#/components/parameters/Scope" }]
      responses:
        "200": { $ref: "#/components/responses/List" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [Tasks]
      operationId: createTask
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses:
        "201": { $ref: "#/components/responses/Entity" }
        default: { $ref: "#/components/responses/Error" }
  /v1/seniors/{seniorId}/tasks/templates:
    parameters: [{ $ref: "#/components/parameters/SeniorId" }]
    get:
      tags: [Tasks]
      operationId: listTaskTemplates
      responses:
        "200": { $ref: "#/components/responses/List" }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/tasks/{taskId}:
    parameters: [{ name: taskId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }]
    get:
      tags: [Tasks]
      operationId: getTask
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
    patch:
      tags: [Tasks]
      operationId: updateTask
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
    delete:
      tags: [Tasks]
      operationId: cancelTask
      responses: { "204": { description: Cancelled }, default: { $ref: "#/components/responses/Error" } }
  /v1/tasks/{taskId}/complete:
    parameters: [{ name: taskId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }]
    post:
      tags: [Tasks]
      operationId: completeTask
      requestBody: { $ref: "#/components/requestBodies/OptionalJsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/tasks/{taskId}/skip:
    parameters: [{ name: taskId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }]
    post:
      tags: [Tasks]
      operationId: skipTask
      requestBody: { $ref: "#/components/requestBodies/OptionalJsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    SeniorId: { name: seniorId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
    Scope: { name: scope, in: query, schema: { type: string } }
  requestBodies:
    JsonObject:
      required: true
      content: { application/json: { schema: { type: object } } }
    OptionalJsonObject:
      required: false
      content: { application/json: { schema: { type: object } } }
...
  responses:
...
    Entity:
      description: Resource
      content: { application/json: { schema: { type: object } } }
    List:
      description: Resource list
      content: { application/json: { schema: { $ref: "#/components/schemas/ListEnvelope" } } }
...
    Error:
      description: Request failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
    Uuid: { type: string, format: uuid }
    ListEnvelope:
      type: object
      required: [items]
      properties: { items: { type: array, items: { type: object } } }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: { type: string } }